Split file model settings and serialization mixins.
